+++ src/upload/mod.rs
@@ -1,8 +1,6 @@
 //! The upload module provides the package upload functionality.
 
-use crate::{
-    tool_configuration::APP_USER_AGENT, AnacondaData, ArtifactoryData, PrefixData, QuetzData,
-};
+use crate::{tool_configuration::APP_USER_AGENT, AnacondaData, ArtifactoryData, QuetzData};
 use fs_err::tokio as fs;
 use futures::TryStreamExt;
 use indicatif::{style::TemplateError, HumanBytes, ProgressState};
@@ -13,15 +11,11 @@ use std::{
     time::{Duration, SystemTime},
 };
 use tokio_util::io::ReaderStream;
-use trusted_publishing::{check_trusted_publishing, TrustedPublishResult};
 
 use miette::{Context, IntoDiagnostic};
 use rattler_networking::{Authentication, AuthenticationStorage};
 use rattler_redaction::Redact;
-use reqwest::{
-    header::{self, HeaderMap, HeaderValue},
-    Method, StatusCode,
-};
+use reqwest::{Method, StatusCode};
 use tracing::{info, warn};
 use url::Url;
 
@@ -30,7 +24,10 @@ use crate::upload::package::{sha256_sum, ExtractedPackage};
 mod anaconda;
 pub mod conda_forge;
 mod package;
+mod prefix;
 mod trusted_publishing;
+
+pub use prefix::upload_package_to_prefix;
 
 const VERSION: &str = env!("CARGO_PKG_VERSION");
 
@@ -198,157 +195,6 @@ pub async fn upload_package_to_artifactory(
     }
 
     info!("Packages successfully uploaded to Artifactory server");
-
-    Ok(())
-}
-
-/// Uploads package files to a prefix.dev server.
-pub async fn upload_package_to_prefix(
-    storage: &AuthenticationStorage,
-    package_files: &Vec<PathBuf>,
-    prefix_data: PrefixData,
-) -> miette::Result<()> {
-    let check_storage = || {
-        match storage.get_by_url(Url::from(prefix_data.url.clone())) {
-            Ok((_, Some(Authentication::BearerToken(token)))) => Ok(token),
-            Ok((_, Some(_))) => {
-                Err(miette::miette!("A Conda token is required for authentication with prefix.dev.
-                        Authentication information found in the keychain / auth file, but it was not a Bearer token"))
-            }
-            Ok((_, None)) => {
-                Err(miette::miette!(
-                    "No prefix.dev api key was given and none was found in the keychain / auth file"
-                ))
-            }
-            Err(e) => {
-                Err(miette::miette!(
-                    "Failed to get authentication information from keychain: {e}"
-                ))
-            }
-        }
-    };
-
-    let token = match prefix_data.api_key {
-        Some(api_key) => api_key,
-        None => match check_trusted_publishing(
-            &get_client_with_retry().into_diagnostic()?,
-            &prefix_data.url,
-        )
-        .await
-        {
-            TrustedPublishResult::Configured(token) => token.secret().to_string(),
-            TrustedPublishResult::Skipped => {
-                if prefix_data.attestation.is_some() {
-                    return Err(miette::miette!(
-                        "An attestation was provided, but trusted publishing is not configured"
-                    ));
-                }
-                check_storage()?
-            }
-            TrustedPublishResult::Ignored(err) => {
-                tracing::warn!("Checked for trusted publishing but failed with {err}");
-                if prefix_data.attestation.is_some() {
-                    return Err(miette::miette!(
-                        "An attestation was provided, but trusted publishing is not configured"
-                    ));
-                }
-                check_storage()?
-            }
-        },
-    };
-
-    for package_file in package_files {
-        let filename = package_file
-            .file_name()
-            .expect("no filename found")
-            .to_string_lossy()
-            .to_string();
-
-        let file_size = package_file.metadata().into_diagnostic()?.len();
-
-        let url = prefix_data
-            .url
-            .join(&format!("api/v1/upload/{}", prefix_data.channel))
-            .into_diagnostic()?;
-
-        let mut form = reqwest::multipart::Form::new();
-
-        let progress_bar = indicatif::ProgressBar::new(file_size)
-            .with_prefix("Uploading")
-            .with_style(default_bytes_style().into_diagnostic()?);
-
-        let progress_bar_clone = progress_bar.clone();
-        let reader_stream =
-            ReaderStream::new(fs::File::open(package_file).await.into_diagnostic()?)
-                .inspect_ok(move |bytes| {
-                    progress_bar_clone.inc(bytes.len() as u64);
-                })
-                .inspect_err(|e| {
-                    println!("Error while uploading: {}", e);
-                });
-
-        let hash = sha256_sum(package_file).into_diagnostic()?;
-
-        let mut file_headers = HeaderMap::new();
-        file_headers.insert(
-            header::CONTENT_TYPE,
-            HeaderValue::from_static("application/octet-stream"),
-        );
-        file_headers.insert(
-            header::CONTENT_LENGTH,
-            file_size.to_string().parse().into_diagnostic()?,
-        );
-        file_headers.insert("X-File-Name", filename.parse().unwrap());
-        file_headers.insert("X-File-SHA256", hash.parse().unwrap());
-
-        let file_part = reqwest::multipart::Part::stream_with_length(
-            reqwest::Body::wrap_stream(reader_stream),
-            file_size,
-        )
-        .file_name(filename)
-        .headers(file_headers);
-
-        form = form.part("file", file_part);
-
-        if let Some(attestation) = &prefix_data.attestation {
-            let text = fs::read_to_string(attestation).await.into_diagnostic()?;
-            let attestation = reqwest::multipart::Part::text(text);
-            form = form.part("attestation", attestation);
-        }
-
-        // Note we cannot use the reqwest client with middleware because we stream
-        // the file during upload
-        let prepared_request = get_default_client()
-            .into_diagnostic()?
-            .post(url.clone())
-            .multipart(form)
-            .bearer_auth(token.clone());
-
-        // send_request_with_retry(prepared_request, package_file).await?;
-        let response = prepared_request.send().await.into_diagnostic()?;
-
-        if response.status().is_success() {
-            progress_bar.finish();
-            info!(
-                "Upload complete for package file: {}",
-                package_file
-                    .file_name()
-                    .expect("no filename found")
-                    .to_string_lossy()
-            );
-        } else {
-            let status = response.status();
-            let body = response.text().await.into_diagnostic()?;
-            return Err(miette::miette!(
-                "Failed to upload package file: {}\nStatus: {}\nBody: {}",
-                package_file.display(),
-                status,
-                body
-            ));
-        }
-    }
-
-    info!("Packages successfully uploaded to prefix.dev server");
 
     Ok(())
 }
